+++ e2e/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @cipherstash/e2e
 
+## 0.0.3-rc.0
+
+### Patch Changes
+
+- Updated dependencies [31ca318]
+- Updated dependencies [31ca318]
+- Updated dependencies [229ce59]
+- Updated dependencies [c4787c0]
+- Updated dependencies [66a0e02]
+- Updated dependencies [cfd46ee]
+- Updated dependencies [0b9b192]
+- Updated dependencies [7eba32d]
+- Updated dependencies [0ebf57e]
+- Updated dependencies [d73a03c]
+- Updated dependencies [89b903f]
+- Updated dependencies [229ce59]
+- Updated dependencies [82f2e69]
+- Updated dependencies [50c0a9c]
+- Updated dependencies [63ca540]
+- Updated dependencies [5d23e80]
+- Updated dependencies [1aa9a11]
+- Updated dependencies [af2d04e]
+- Updated dependencies [b8a3d20]
+- Updated dependencies [a0f3b2c]
+- Updated dependencies [f23f952]
+- Updated dependencies [0b9b192]
+- Updated dependencies [7c7dbca]
+- Updated dependencies [5411a13]
+- Updated dependencies [e25eb22]
+- Updated dependencies [1a9d190]
+- Updated dependencies [161f17b]
+- Updated dependencies [e40c3da]
+- Updated dependencies [58d7439]
+- Updated dependencies [99f8b0a]
+- Updated dependencies [fd33aad]
+- Updated dependencies [8cd485d]
+- Updated dependencies [9b65ae8]
+- Updated dependencies [9c673bb]
+  - @cipherstash/stack@1.0.0-rc.0
+  - stash@1.0.0-rc.0
+  - @cipherstash/wizard@0.5.0-rc.0
+  - @cipherstash/drizzle@3.0.4-rc.0
+  - @cipherstash/protect@12.0.2-rc.0
+
 ## 0.0.2
 
 ### Patch Changes